requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
supabase>=2.0.0
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
pypdf>=4.0.0
tiktoken>=0.7.0
numpy>=1.26.0
